move ProblemException to utils
